backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.24.1
supabase==2.3.4
groq==0.8.0
python-dotenv==1.0.1
pydantic==2.7.1
youtube-transcript-api==0.6.2
yt-dlp==2024.4.9
pyPDF==4.2.0
python-pptx==0.6.23
aiofiles==23.2.1
